requirements.txt
psycopg2
Pillow
django-crispy-forms